update turbine tank
